SERVER-7886 - use StringData methods, now .data()

+++ src/mongo/db/dur_recover.cpp
@@ -259,7 +259,7 @@ namespace mongo {
             //TODO(mathias): look into making some of these dasserts
             verify(entry.e);
             verify(entry.dbName);
-            verify(strnlen(entry.dbName, MaxDatabaseNameLen) < MaxDatabaseNameLen);
+            verify((size_t)strnlen(entry.dbName, MaxDatabaseNameLen) < MaxDatabaseNameLen);
 
             MongoMMF *mmf = last.newEntry(entry, *this);
 
@@ -309,7 +309,7 @@ namespace mongo {
 
         MongoMMF* RecoveryJob::getMongoMMF(const ParsedJournalEntry& entry) {
             verify(entry.dbName);
-            verify(strnlen(entry.dbName, MaxDatabaseNameLen) < MaxDatabaseNameLen);
+            verify((size_t)strnlen(entry.dbName, MaxDatabaseNameLen) < MaxDatabaseNameLen);
 
             const string fn = fileName(entry.dbName, entry.e->getFileNo());
             MongoFile* file;
